Various improvements
1. Add support to compute morton code
2. Make octree support float and double
3. Make sensor simulation support float and double: lidar, depth, rgbd
4. Make some util functions support float and double
5. Improve factory pattern

+++ include/erl_geometry/logodd.hpp
@@ -5,11 +5,11 @@
 namespace erl::geometry::logodd {
 
     inline float
-    LogOdd(const double p) {
-        return static_cast<float>(std::log(p / (1 - p)));
+    LogOdd(const float p) {
+        return std::log(p / (1 - p));
     }
 
-    inline double
+    inline float
     Probability(const float logodd) {
         return 1.0 / (1.0 + std::exp(-logodd));
     }
